Update integration/aws-func to use the new application-tester package
