+++ packages/framework-integration-tests/integration/providers/aws/functionality/setup.ts
@@ -1,6 +1,9 @@
-import { setEnv, checkConfigAnd, hopeTheBest } from '../utils'
+import { ApplicationTester } from '@boostercloud/application-tester'
+import { getProviderTestHelper, setEnv } from '../../../helper/app-helper'
+
+export let applicationUnderTest: ApplicationTester
 
 before(async () => {
   await setEnv()
-  await checkConfigAnd(hopeTheBest)
+  applicationUnderTest = new ApplicationTester(await getProviderTestHelper())
 })
